[fix] 인터페이스 제거, spring data jpa 사용, Optional 사용
