+++ football_love/src/main/java/com/deu/football_love/controller/ExceptionController.java
@@ -1,6 +1,7 @@
 package com.deu.football_love.controller;
 
 import com.deu.football_love.exception.LikeDuplicatedException;
+import com.deu.football_love.exception.NotTeamMemberException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,9 +15,17 @@ public class ExceptionController {
     @ExceptionHandler({
             LikeDuplicatedException.class,
     })
-    public ResponseEntity<Object> LikeDuplicatedException(final RuntimeException ex) {
+    public ResponseEntity<Object> likeDuplicatedException(final RuntimeException ex) {
         log.warn("error", ex);
         return new ResponseEntity<>(HttpStatus.CONFLICT);
+    }
+
+    @ExceptionHandler({
+            NotTeamMemberException.class,
+    })
+    public ResponseEntity<Object> notTeamMemberException(final RuntimeException ex) {
+        log.warn("error", ex);
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
