+++ football_love/src/main/java/com/deu/football_love/exception/NotTeamMemberException.java
@@ -1,0 +1,10 @@
+package com.deu.football_love.exception;
+
+public class NotTeamMemberException  extends RuntimeException{
+    public NotTeamMemberException(String message) {
+        super(message);
+    }
+
+    public NotTeamMemberException() {
+    }
+}
