+++ football_love/src/main/java/com/deu/football_love/repository/StadiumRepository.java
@@ -1,14 +1,15 @@
 package com.deu.football_love.repository;
 
 import com.deu.football_love.domain.Stadium;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface StadiumRepository {
-    Stadium selectStadium(Long stadiumId);
+public interface StadiumRepository extends JpaRepository<Stadium, Long> {
 
-    List<Stadium> selectAllStadiumByCompanyId(Long companyId);
+    @Query("SELECT s FROM Stadium s WHERE s.company.id = :companyId")
+    List<Stadium> selectAllStadiumByCompanyId(@Param("companyId") Long companyId);
 
-    void insertStadium(Stadium stadium);
-    void deleteStadium(Stadium stadium);
 }
