+++ football_love/src/main/java/com/deu/football_love/repository/TeamRepository.java
@@ -3,20 +3,15 @@ package com.deu.football_love.repository;
 import com.deu.football_love.domain.ApplicationJoinTeam;
 import com.deu.football_love.domain.Team;
 import com.deu.football_love.domain.TeamMember;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface TeamRepository {
+public interface TeamRepository extends JpaRepository<Team, Long> {
 
-    void insertTeam(Team team);
-    Team selectTeam(Long teamId);
-    Team selectTeamByName(String teamName);
-    void insertNewApplication(ApplicationJoinTeam application);
-    TeamMember insertNewTeamMember(TeamMember newTeamMember);
-    ApplicationJoinTeam selectApplication(Long teamId, String memberId);
-    void deleteApplication(ApplicationJoinTeam application);
-    void deleteTeamMember(Long teamId, Long memberNumber);
-    List<TeamMember> selectTeamMember(Long teamId, Long memberNumber);
-    void deleteTeam(Team team);
-    boolean existsTeamByTeamName(String teamName);
+    Optional<Team> findByName(@Param("name") String teamName);
+    boolean existsTeamByName(@Param("name") String teamName);
+
 }
