+++ football_love/src/main/java/com/deu/football_love/service/TeamService.java
@@ -1,35 +1,170 @@
 package com.deu.football_love.service;
 
+import com.deu.football_love.domain.*;
 import com.deu.football_love.domain.type.TeamMemberType;
 import com.deu.football_love.dto.team.*;
+import com.deu.football_love.exception.NotTeamMemberException;
+import com.deu.football_love.repository.*;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
-public interface TeamService {
+@Service
+@RequiredArgsConstructor
+@Slf4j
+@Transactional
+public class TeamService {
 
-    QueryTeamDto getTeamInfo(Long teamId);
+    private final TeamRepository teamRepository;
+    private final MemberRepository memberRepository;
+    private final PostRepository postRepository;
+    private final BoardRepository boardRepository;
+    private final TeamMemberRepository teamMemberRepository;
+    private final ApplicationJoinTeamRepository applicationJoinTeamRepository;
 
-    CreateTeamResponse createNewTeam(String creator, String teamName);
+    @Transactional(readOnly = true)
+    public QueryTeamDto getTeamInfo(Long teamId) {
+        Team findTeam = teamRepository.findById(teamId).orElseThrow(()->new IllegalArgumentException("no such Team data"));
+        return QueryTeamDto.from(findTeam);
+    }
 
-    QueryTeamDto findTeam(Long teamId);
+    public CreateTeamResponse createNewTeam(String creator, String teamName) {
+        Team team = new Team();
+        Member findMember = memberRepository.selectMemberById(creator);
+        if (findMember == null)
+            return null;
+        TeamMember teamMember = new TeamMember(team, findMember, TeamMemberType.LEADER);
+        team.setName(teamName);
+        teamMember.setTeam(team);
+        teamMember.setMember(findMember);
+        team.getTeamMembers().add(teamMember);
+        findMember.getTeamMembers().add(teamMember);
+        teamRepository.save(team);
+        teamMemberRepository.save(teamMember);
+        return new CreateTeamResponse(team.getId());
+    }
 
-    QueryTeamDto findTeamByName(String teamName);
+    @Transactional(readOnly = true)
+    public QueryTeamDto findTeam(Long teamId) {
+        Team findTeam = teamRepository.findById(teamId).orElseThrow(()->new IllegalArgumentException("no such Team data"));
+        return QueryTeamDto.from(findTeam);
+    }
 
-    ApplicationJoinTeamDto findApplication(Long teamId, String memberId);
+    @Transactional(readOnly = true)
+    public QueryTeamDto findTeamByName(String teamName) {
+        Team team = teamRepository.findByName(teamName).orElseThrow(()->new IllegalArgumentException("no such Team data"));
+        return QueryTeamDto.from(team);
+    }
 
-    void applyToTeam(Long teamId, String memberId, String message);
+    @Transactional(readOnly = true)
+    public ApplicationJoinTeamDto findApplication(Long teamId, String memberId) {
+        ApplicationJoinTeam application = applicationJoinTeamRepository.findByTeamIdAndMemberId(teamId, memberId)
+                .orElseThrow(()->new IllegalArgumentException("no such Team data"));
+        if (application == null)
+            return null;
+        return new ApplicationJoinTeamDto(application);
+    }
 
-    AcceptApplicationResponse acceptApplication(Long teamId, String memberId);
+    public void applyToTeam(Long teamId, String memberId, String message) {
 
-    TeamMemberType authorityCheck(Long teamId, Long memberNumber);
+        Team findTeam = teamRepository.findById(teamId).orElseThrow(()->new IllegalArgumentException("no such Team data"));
+        Member findMember = memberRepository.selectMemberById(memberId);
 
-    void withdrawal(Long teamId, String memberId);
+        ApplicationJoinTeam application = new ApplicationJoinTeam(findTeam, findMember, message);
+        findTeam.getApplicationJoinTeams().add(application);
+        findMember.getApplicationJoinTeams().add(application);
 
-    DisbandmentTeamResponse disbandmentTeam(Long teamId);
+        applicationJoinTeamRepository.save(application);
+    }
 
-    List<QueryTeamMemberDto> findTeamMember(Long teamId, Long memberNumber);
+    public AcceptApplicationResponse acceptApplication(Long teamId, String memberId) {
+        ApplicationJoinTeam findApplication = applicationJoinTeamRepository.findByTeamIdAndMemberId(teamId, memberId)
+                .orElseThrow(()->new IllegalArgumentException("no such Team data"));;
+        Team findTeam = teamRepository.findById(teamId).orElseThrow(()->new IllegalArgumentException("no such Team data"));
+        Member findMember = memberRepository.selectMemberById(memberId);
+        if (findMember == null)
+            return null;
+        findApplication.delete();
+        applicationJoinTeamRepository.delete(findApplication);
+        TeamMember newTeamMember = new TeamMember(findTeam, findMember, TeamMemberType.COMMON);
+        findTeam.getTeamMembers().add(newTeamMember);
+        findMember.getTeamMembers().add(newTeamMember);
+        teamMemberRepository.save(newTeamMember);
 
-    List<QueryTeamMemberDto> findTeamMemberByMemberId(Long teamId, String memberId);
+        return new AcceptApplicationResponse(teamId, memberId);
+    }
 
-    UpdateAuthorityResponse updateAuthority(Long teamId, String memberId, TeamMemberType authorityType);
+    /**
+     * 팀에 속해 있는지, 속해 있다면 일반 회원인지 관리자인지 파악
+     */
+    @Transactional(readOnly = true)
+    public TeamMemberType authorityCheck(Long teamId, Long memberNumber) {
+        Optional<TeamMember> findTeamMember = teamMemberRepository.findByTeamIdAndMemberNumber(teamId, memberNumber);
+        if (!findTeamMember.isPresent())
+            return TeamMemberType.NONE;
+        return findTeamMember.get().getType();
+    }
+
+    @Transactional(readOnly = true)
+    public QueryTeamMemberDto findTeamMemberByTeamIdAndMemberNumber(Long teamId, Long memberNumber) {
+        TeamMember teamMember = teamMemberRepository.findByTeamIdAndMemberNumber(teamId, memberNumber)
+                .orElseThrow(() -> new NotTeamMemberException("no such team_member data"));
+        return new QueryTeamMemberDto(teamMember);
+    }
+
+    @Transactional(readOnly = true)
+    public QueryTeamMemberDto findTeamMemberByTeamIdAndMemberId(Long teamId, String memberId) {
+        TeamMember teamMember = teamMemberRepository.findByTeamIdAndMemberId(teamId, memberId)
+                .orElseThrow(() -> new NotTeamMemberException("no such team_member data"));
+        return new QueryTeamMemberDto(teamMember);
+    }
+
+    @Transactional(readOnly = true)
+    public List<QueryTeamMemberDto> findTeamMembersByTeamId(Long teamId) {
+        return teamMemberRepository.findTeamMembersByTeamId(teamId).stream().map(teamMember -> new QueryTeamMemberDto(teamMember)).collect(Collectors.toList());
+    }
+
+    public void withdrawal(Long teamId, String memberId) {
+        Member findMember = memberRepository.selectMemberById(memberId);
+        if(!teamMemberRepository.existsByTeamIdAndMemberId(teamId, memberId))
+            throw new NotTeamMemberException("no such team_member data");
+        teamMemberRepository.deleteByTeamIdAndMemberNumber(teamId, findMember.getNumber());
+    }
+
+    public DisbandmentTeamResponse disbandmentTeam(Long teamId) {
+        Team findTeam = teamRepository.findById(teamId).orElseThrow(()->new IllegalArgumentException("no such Team data"));
+        while (findTeam.getBoards().size() != 0) {
+            Board board = findTeam.getBoards().get(0);
+            while (board.getPosts().size() != 0) {
+                Post post = board.getPosts().get(0);
+                post.deletePost();
+                postRepository.delete(post);
+            }
+            board.getTeam().getBoards().remove(board);
+            board.setTeam(null);
+            boardRepository.delete(board);
+        }
+        while (findTeam.getTeamMembers().size() != 0) {
+            TeamMember teamMember = findTeam.getTeamMembers().get(0);
+            Long teamMemberId = teamMember.getId();
+            teamMember.deleteTeamMember();;
+            teamMemberRepository.deleteByTeamIdAndMemberNumber(teamId, teamMemberId);
+        }
+        teamRepository.delete(findTeam);
+        return new DisbandmentTeamResponse(teamId);
+    }
+
+    public UpdateAuthorityResponse updateAuthority(Long teamId, String memberId, TeamMemberType authorityType) {
+        Member findMember = memberRepository.selectMemberById(memberId);
+        TeamMember teamMember = teamMemberRepository.findByTeamIdAndMemberNumber(teamId, findMember.getNumber())
+                .orElseThrow(() -> new IllegalArgumentException("no such team_member data"));
+        teamMember.setType(authorityType);
+        return new UpdateAuthorityResponse(teamId, memberId, authorityType);
+    }
+
 }
